+++ tests/test_keep_alive.py
@@ -1,49 +1,71 @@
-
 import pytest
 import asyncio
-import websockets
+from roll import Protocol
+from roll.testing import Transport
+
+
+@pytest.fixture
+def protocol(app, event_loop):
+    app.loop = event_loop
+    protocol = Protocol(app)
+    protocol.connection_made(Transport())
+    return protocol
 
 
 @pytest.mark.asyncio
-async def test_keep_alive(liveclient):
+async def test_not_persistent(protocol):
+    protocol.data_received(
+        b'GET /test HTTP/1.1\r\n'
+        b'Host: localhost:1707\r\n'
+        b'User-Agent: Lynx\r\n'
+        b'Connection: Close\r\n'
+        b'\r\n')
+    assert protocol.keep_alive == False
+    assert not protocol.watchers
 
-    @liveclient.app.route('/hello')
-    async def hello(request, response):
-        response.body = b'Hello !'
 
-    @liveclient.app.route('/world')
-    async def world(request, response):
-        response.body = b'World !'
+@pytest.mark.asyncio
+async def test_switch(protocol):
 
-    @liveclient.app.route('/long')
-    async def long_task(request, response):
-        # The keep alive is in between requests, no during one.
-        # So sleeping over the timeout time is of no consequence.
-        await asyncio.sleep(11)
-        response.body = b'I did a long task !'
+    # We keep the connection alive after the first request
+    protocol.data_received(
+        b'GET /test HTTP/1.1\r\n'
+        b'Host: localhost:1707\r\n'
+        b'User-Agent: Lynx\r\n'
+        b'Connection: Keep-Alive\r\n'
+        b'\r\n')
+    assert protocol.keep_alive == True
+    assert not protocol.watchers
 
-    with liveclient as query:
-        response = await query('GET', '/hello')
-        assert response.status == 200
-        assert response.read() == b'Hello !'
+    # Replying here will create the keep alive timeout task
+    # We simulate this by calling "prime_for_request", as the
+    # reply method does after writing
+    protocol.prime_for_request()
+    assert protocol.watchers['keep_alive']
+    
+    # We ask for close for the next request
+    protocol.data_received(
+        b'GET /test HTTP/1.1\r\n'
+        b'Host: localhost:1707\r\n'
+        b'User-Agent: Lynx\r\n'
+        b'Connection: Close\r\n'
+        b'\r\n')
+    assert protocol.keep_alive == False
 
-        response = await query('GET', '/world')
-        assert response.status == 200
-        assert response.read() == b'World !'
+    # The timeout task was cleared
+    assert not protocol.watchers
 
-        response = await query('GET', '/unknown')
-        assert response.status == 404
-        assert response.read() == b'/unknown'
+    # Priming the request again will not create a timeout
+    # task, as we are expected to close the connection.
+    protocol.prime_for_request()
+    assert not protocol.watchers
 
-        response = await query('GET', '/long')
-        assert response.status == 200
-        assert response.read() == b'I did a long task !'
-        
-        response = await query('GET', '/world')
-        assert response.status == 200
-        assert response.read() == b'World !'
-        
-        await asyncio.sleep(11)  # Currently the timeout is set to 10
-
-        response = await query('GET', '/hello')
-        assert response.status == 408  # Request Timeout
+    # If we ask again, we are expected to be closed, so new keep_alive
+    # Won't mean a thing, until we re-create a protocol.
+    protocol.data_received(
+        b'GET /test HTTP/1.1\r\n'
+        b'Host: localhost:1707\r\n'
+        b'User-Agent: Lynx\r\n'
+        b'Connection: Keep-Alive\r\n'
+        b'\r\n')
+    assert protocol.keep_alive == False
